+++ rest_framework/filters.py
@@ -4,7 +4,7 @@ returned by list views.
 """
 from __future__ import unicode_literals
 from django.db import models
-from rest_framework.compat import django_filters, six, guardian
+from rest_framework.compat import django_filters, six, guardian, get_model_name
 from functools import reduce
 import operator
 
@@ -124,6 +124,7 @@ class OrderingFilter(BaseFilterBackend):
 
     def remove_invalid_fields(self, queryset, ordering):
         field_names = [field.name for field in queryset.model._meta.fields]
+        field_names += queryset.query.aggregates.keys()
         return [term for term in ordering if term.lstrip('-') in field_names]
 
     def filter_queryset(self, request, queryset, view):
@@ -158,7 +159,7 @@ class DjangoObjectPermissionsFilter(BaseFilterBackend):
         model_cls = queryset.model
         kwargs = {
             'app_label': model_cls._meta.app_label,
-            'model_name': model_cls._meta.module_name
+            'model_name': get_model_name(model_cls)
         }
         permission = self.perm_format % kwargs
         return guardian.shortcuts.get_objects_for_user(user, permission, queryset)